Major overhaul to use backstretch
I decided to use backstretch.js to make backgrounds fit on all devices
as well as pre-load all images to prevent jittering load-in.

CURRENT ISSUES:
1. If the quote extends beyond screen on iOS devices, gray background
will show when you scroll down to read all of the quote. This is a
known issue on iOS and there isn’t a way to fix it to my knowledge.
2. On mobile, occasionally it takes a while for the next background to
load but the current background will still display until the next one
is ready.
3. On desktop and possibly on mobile, sometimes the fade in for the
next photo/quote is jittery. This may be a limitation on jquery or
because backstretch.js is trying to fade as well.
Note: Some of these issues may “fix” themselves once images have been
cached in the browser but I’m not 100% sure.